+++ src/main/java/top/ellan/mahjong/bootstrap/MahjongPaperLoader.java
@@ -18,12 +18,12 @@ public final class MahjongPaperLoader implements PluginLoader {
         ).build());
 
         this.addDependency(resolver, "io.github.ssttkkl:mahjong-utils-jvm:0.7.7");
-        this.addDependency(resolver, "org.mariadb.jdbc:mariadb-java-client:3.5.3");
-        this.addDependency(resolver, "com.mysql:mysql-connector-j:8.4.0");
-        this.addDependency(resolver, "com.h2database:h2:2.3.232");
-        this.addDependency(resolver, "com.zaxxer:HikariCP:6.3.0");
-        this.addDependency(resolver, "org.jetbrains.kotlin:kotlin-stdlib:2.2.0");
-        this.addDependency(resolver, "org.jetbrains.kotlinx:kotlinx-serialization-json:1.9.0");
+        this.addDependency(resolver, "org.mariadb.jdbc:mariadb-java-client:3.5.9");
+        this.addDependency(resolver, "com.mysql:mysql-connector-j:9.7.0");
+        this.addDependency(resolver, "com.h2database:h2:2.4.240");
+        this.addDependency(resolver, "com.zaxxer:HikariCP:7.0.2");
+        this.addDependency(resolver, "org.jetbrains.kotlin:kotlin-stdlib:2.4.0");
+        this.addDependency(resolver, "org.jetbrains.kotlinx:kotlinx-serialization-json:1.11.0");
 
         classpathBuilder.addLibrary(resolver);
     }